requirements.txt
basemap==1.1.0
Cerberus==0.9.2
certifi==2018.1.18
chardet==3.0.4
click==6.7
colorspacious==1.1.0
cycler==0.10.0
Cython==0.27.3
decorator==4.2.1
Django==2.0.1
docopt==0.6.2
et-xmlfile==1.0.1
Eve==0.7.6
Events==0.2.2
Flask==0.12.2
Flask-MySQL==1.4.0
flask-paginate==0.5.1
Flask-PyMongo==0.5.1
Flask-Restless==0.17.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
future==0.16.0
gmplot==1.2.0
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
itsdangerous==0.24
jdcal==1.3
jedi==0.11.1
Jinja2==2.10
MarkupSafe==1.0
matplotlib==2.1.2
mimerender==0.6.0
mpld3==0.3
mysqlclient==1.3.12
networkx==2.1
nltk==3.2.5
numpy==1.14.0
oauthlib==2.0.6
odfpy==1.3.6
openpyxl==2.5.1
pandas==0.22.0
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==5.0.0
pkg-resources==0.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pygal==2.4.0
Pygments==2.2.0
pymongo==3.6.0
PyMySQL==0.8.0
pyparsing==2.2.0
pyproj==1.9.5.1
PySAL==1.14.3
python-dateutil==2.6.1
python-highcharts==0.4.1
python-mimeparse==1.6.0
pytz==2017.3
PyYAML==3.12
records==0.5.2
requests==2.18.4
requests-oauthlib==0.8.0
scipy==1.0.0
Shapely==1.6.4.post1
simplegeneric==0.8.1
simplejson==3.13.2
six==1.11.0
SQLAlchemy==1.2.1
tablib==0.12.1
textblob==0.15.1
traitlets==4.3.2
twython==3.6.0
unicodecsv==0.14.1
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.14.1
wordcloud==1.4.1
WTForms==2.1
xlrd==1.1.0
xlwt==1.3.0
